+++ src/JsonRpc/JsonRpcClient.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2015-2018, The Bytecoin (Byterub) developers.
+// Copyright (c) 2015-2018, The Bytecoin developers.
 // Licensed under the GNU Lesser General Public License. See LICENSE for details.
 
 #pragma once
@@ -19,7 +19,7 @@
 #include "JsonRpcNotification.h"
 #include "JsonRpcObjectFactory.h"
 #include "rpcapi.h"
-//#include "byterub/rpc_api.h"
+//#include "bytecoin/rpc_api.h"
 
 namespace JsonRpc {