+++ src/walletmodel.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2015-2018, The Bytecoin (Byterub) developers.
+// Copyright (c) 2015-2018, The Bytecoin developers.
 // Licensed under the GNU Lesser General Public License. See LICENSE for details.
 
 #include <QLocale>
@@ -649,7 +649,7 @@ QVariant WalletModel::getDisplayRoleStatus(const QModelIndex& index) const
     }
     case COLUMN_LOWER_LEVEL_ERROR:
 //        return pimpl_->status.lower_level_error;
-        return pimpl_->status.lower_level_error.isEmpty() ? tr("Connected to byterubd") : tr("Byterubd status: %1").arg(pimpl_->status.lower_level_error);
+        return pimpl_->status.lower_level_error.isEmpty() ? tr("Connected to bytecoind") : tr("Bytecoind status: %1").arg(pimpl_->status.lower_level_error);
     case COLUMN_NEXT_BLOCK_EFFECTIVE_MEDIAN_SIZE:
         return pimpl_->status.next_block_effective_median_size;
     case COLUMN_TXPOOL_VERSION: