+++ ReleaseNotes.md
@@ -5,7 +5,7 @@
 - Added creating send proofs.
 - Added checking send proofs.
 - Added wallet daemon parameters window.
-- Added byterubd status in the status bar.
+- Added bytecoind status in the status bar.
 
 ### v2.0.0-beta-20180219